utils: Extract pkg meta data utilities into separate module
